Handle being already signed in better
When already signed in, unnecessary auth calls are made to Google API. Already being signed in is now handled better.

This also gets rid of a weird flash when already signed in and making calls to `loginIfNeeded()`

Also api setup issues were hard to debug because it was not logging the exception code from google. This is also fixed.